6.3.5

+++ gc-includes/class-phpass.php
@@ -2,7 +2,6 @@
 /**
  * Portable PHP password hashing framework.
  * @package phpass
- *
  * @version 0.5 / GeChiUI
  * @link https://www.openwall.com/phpass/
  */
@@ -39,7 +38,6 @@
  * @package phpass
  * @version 0.5 / GeChiUI
  * @link https://www.openwall.com/phpass/
- *
  */
 class PasswordHash {
 	var $itoa64;
@@ -173,8 +171,8 @@ class PasswordHash {
 		$itoa64 = './ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
 
 		$output = '$2a$';
-		$output .= chr(ord('0') + $this->iteration_count_log2 / 10);
-		$output .= chr(ord('0') + $this->iteration_count_log2 % 10);
+		$output .= chr((int)(ord('0') + $this->iteration_count_log2 / 10));
+		$output .= chr((ord('0') + $this->iteration_count_log2 % 10));
 		$output .= '$';
 
 		$i = 0;
